project/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
boto3==1.17.68
botocore==1.20.68
celery==5.0.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary
cryptography==2.9.2
Django==3.2
django-cors-headers==3.7.0
django-polymorphic
django-rest-polymorphic
djangorestframework==3.12.4
djangorestframework-simplejwt
drf-nested-routers==0.93.3
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.0.1
jmespath==0.10.0
kombu==5.0.2
openai
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
pycurl==7.43.0.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==6.0
requests==2.26.0
requests-aws4auth==1.1.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
stripe
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
channels==3.0.4